requirements.txt
PyQt6>=6.6.0
python-docx>=1.1.0
reportlab>=4.1.0
Pillow>=10.3.0
opencv-python>=4.9.0
requests>=2.31.0
